# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  backend_internal_httpresp.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  backend_internal_httpresp.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/backend_internal_httpresp.ErrorBody'
    type: object
...
  internal_jobs.JobListResponse:
    properties:
      jobs:
        items:
          $ref: '#/definitions/internal_jobs.JobResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 42
        type: integer
    type: object
  internal_jobs.JobResponse:
    properties:
      application_url:
        example: https://boards.greenhouse.io/openai/jobs/12345
        type: string
      company:
        example: OpenAI
        type: string
      company_url:
        example: https://openai.com
        type: string
      created_at:
        example: "2026-01-15T08:00:00Z"
        type: string
      description:
        example: We are looking for a Senior Go Engineer...
        type: string
      external_id:
        example: gh-12345
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      location:
        example: San Francisco, CA (Remote)
        type: string
      match_details:
        type: object
      match_score:
        example: 92.5
        type: number
      posted_at:
        example: "2026-01-10T10:00:00Z"
        type: string
      remote_type:
        enum:
        - remote
        - hybrid
        - onsite
        example: remote
        type: string
      requirements:
        example: 5+ years Go experience, Kubernetes, PostgreSQL
        type: string
      salary_currency:
        example: USD
        type: string
      salary_max:
        example: 200000
        type: integer
      salary_min:
        example: 150000
        type: integer
      scraped_at:
        example: "2026-01-15T08:00:00Z"
        type: string
      source:
        example: greenhouse
        type: string
      source_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      source_name:
        example: Greenhouse - OpenAI
        type: string
      status:
        enum:
        - discovered
        - matched
        - applied
        - archived
        example: matched
        type: string
      title:
        example: Senior Go Engineer
        type: string
      updated_at:
        example: "2026-01-15T08:00:00Z"
        type: string
      url:
        example: https://boards.greenhouse.io/openai/jobs/12345
        type: string
    type: object
  internal_jobs.UpdateJobRequest:
    properties:
      status:
        enum:
        - discovered
        - matched
        - applied
        - archived
        example: applied
        type: string
    type: object
  internal_jobs.scanRequest:
    properties:
      source_ids:
        example:
        - '["550e8400-e29b-41d4-a716-446655440000"]'
        items:
          type: string
        minItems: 1
        type: array
    required:
    - source_ids
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/your-org/ai-job-search-agent/issues
  description: Local-first AI-powered job search automation platform. Discovers, scores,
    and applies to jobs with tailored resumes and cover letters.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/your-org/ai-job-search-agent
  title: AI Job Search Agent API
  version: "1.0"
paths:
...
  /job-discovery/scan:
    post:
      consumes:
      - application/json
      description: Start asynchronous job discovery from configured sources
      parameters:
      - description: Source IDs to scan
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_jobs.scanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task IDs for polling
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "202":
          description: Partial failure - some tasks dispatched
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Trigger job discovery scan
      tags:
      - Jobs
  /jobs:
    get:
      consumes:
      - application/json
      description: Get paginated list of jobs with optional filters
      parameters:
      - description: Filter by status
        enum:
        - discovered
        - matched
        - applied
        - archived
        in: query
        name: status
        type: string
      - description: Filter by company name
        in: query
        name: company
        type: string
      - description: Filter by source UUID
        in: query
        name: source_id
        type: string
      - description: Minimum match score (0-100)
        in: query
        maximum: 100
        minimum: 0
        name: min_score
        type: number
      - default: 20
        description: Results per page (max 100)
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/internal_jobs.JobListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List jobs
      tags:
      - Jobs
  /jobs/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific job
      parameters:
      - description: Job UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job details
          schema:
            $ref: '#/definitions/internal_jobs.JobResponse'
        "400":
          description: Invalid job ID
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get job by ID
      tags:
      - Jobs
    patch:
      consumes:
      - application/json
      description: Update the status of a job (e.g., mark as applied, archived)
      parameters:
      - description: Job UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_jobs.UpdateJobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Job updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or status
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_internal_httpresp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update job status
      tags:
      - Jobs
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...